lab_3/src/pyproject.toml
[project]
name = "fastapi-celery"
version = "0.1.0"
description = ""
authors = [
    {name = "couldlie",email = "kudlay@fort-telecom.ru"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.8,<0.116.0)",
    "uvicorn[standard] (>=0.34.0,<0.35.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "pydantic-settings (>=2.8.0,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "alembic (>=1.14.1,<2.0.0)",
    "celery (>=5.4.0,<6.0.0)",
    "paramiko (>=3.5.1,<4.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "flower (>=2.0.1,<3.0.0)",
    "pyserial (>=3.5,<4.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
]

[tool.poetry]
packages = [{include = "fastapi_celery", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
